Setup the basic types
